refactor(kitchenpos): rename product fixture
