+++ src/test/java/kitchenpos/products/tobe/domain/fixture/ProductFixture.java
@@ -1,4 +1,4 @@
-package kitchenpos;
+package kitchenpos.products.tobe.domain.fixture;
 
 import java.math.BigDecimal;
 
@@ -6,8 +6,8 @@ import kitchenpos.common.domain.DisplayedName;
 import kitchenpos.common.domain.Price;
 import kitchenpos.products.tobe.domain.Product;
 
-public class TobeFixtures {
-	public static Product product(final String name, final long price) {
+public class ProductFixture {
+	public static Product 상품(final String name, final long price) {
 		return new Product(
 			new DisplayedName(name, (text) -> false),
 			new Price(new BigDecimal(price))
